Extract to a separate feature
