+++ src/features/basic/unclipped-tooltips.module.css
@@ -1,0 +1,79 @@
+.hideTooltip {
+  &:before {
+    display: none !important;
+    content: none !important;
+  }
+  &:after {
+    display: none !important;
+    content: none !important;
+  }
+}
+
+.popoverTooltip {
+  position: fixed;
+  box-sizing: content-box;
+  pointer-events: none;
+  z-index: 1000;
+
+  background-color: rgb(38, 38, 38);
+  color: rgb(238, 238, 238);
+  margin: 0;
+  padding: 8px;
+  font-size: 11px;
+  font-weight: normal;
+  line-height: 1.1;
+  width: max-content;
+  min-width: 10px;
+  max-width: 200px;
+  white-space: pre-wrap;
+  text-align: left;
+  border-radius: 4px;
+  border: 2px solid #c48400;
+
+  --rp-tooltip-arrow-offset: 50%;
+  overflow: visible;
+
+  &::after {
+    content: '';
+    position: absolute;
+    width: 0;
+    height: 0;
+    border-style: solid;
+  }
+
+  &[data-rp-tooltip-side='bottom']::after {
+    left: var(--rp-tooltip-arrow-offset);
+    top: -10px;
+    transform: translateX(-50%);
+
+    border-width: 0 6px 8px 6px;
+    border-color: transparent transparent #c48400 transparent;
+  }
+
+  &[data-rp-tooltip-side='top']::after {
+    left: var(--rp-tooltip-arrow-offset);
+    bottom: -10px;
+    transform: translateX(-50%);
+
+    border-width: 8px 6px 0 6px;
+    border-color: #c48400 transparent transparent transparent;
+  }
+
+  &[data-rp-tooltip-side='right']::after {
+    left: -10px;
+    top: var(--rp-tooltip-arrow-offset);
+    transform: translateY(-50%);
+
+    border-width: 6px 8px 6px 0;
+    border-color: transparent #c48400 transparent transparent;
+  }
+
+  &[data-rp-tooltip-side='left']::after {
+    right: -10px;
+    top: var(--rp-tooltip-arrow-offset);
+    transform: translateY(-50%);
+
+    border-width: 6px 0 6px 8px;
+    border-color: transparent transparent transparent #c48400;
+  }
+}
